requirements.txt
python-dotenv
requests
kafka-python
pyarrow
pyspark